requirements.txt
torch
diffusers
gradio
transformers
Pillow
einops
lpips
open_clip_torch
tqdm
matplotlib
